+++ eras/allegra/impl/src/Cardano/Ledger/Allegra.hs
@@ -29,16 +29,10 @@ import Cardano.Ledger.Allegra.Tx (Tx (..))
 import Cardano.Ledger.Allegra.UTxO ()
 import Cardano.Ledger.Binary (DecCBOR, EncCBOR)
 import Cardano.Ledger.Block (EraBlockHeader)
-import Cardano.Ledger.Rules.ValidationMode (lblStatic)
 import Cardano.Ledger.Shelley.API
-import Cardano.Ledger.Shelley.Rules (ledgerPpL)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
-import Cardano.Ledger.State (utxoG)
-import Control.State.Transition.Extended (ValidationPolicy (..))
-import Data.Bifunctor (Bifunctor (first))
 import Data.List.NonEmpty (NonEmpty)
 import GHC.Generics (Generic)
-import Lens.Micro ((^.))
 
 instance ApplyTx AllegraEra where
   newtype ApplyTxError AllegraEra
@@ -48,28 +42,9 @@ instance ApplyTx AllegraEra where
 
   mkStAnnTx _ _ _ _ = id
 
-  internalApplyTxWithValidation validationPolicy globals env state tx =
-    let stAnnTx =
-          mkStAnnTx
-            (epochInfo globals)
-            (systemStart globals)
-            (env ^. ledgerPpL)
-            (state ^. utxoG)
-            tx
-     in first AllegraApplyTxError $
-          ruleApplyTxValidation @"LEDGER" validationPolicy globals env state stAnnTx
+  internalApplyTxWithValidation = defaultApplyTxWithValidation @"LEDGER" AllegraApplyTxError
 
-  internalReapplyValidatedTx globals env state vtx =
-    fst
-      <$> first
-        AllegraApplyTxError
-        ( ruleApplyTxValidation @"LEDGER"
-            (ValidateSuchThat (notElem lblStatic))
-            globals
-            env
-            state
-            (vtStAnnTx vtx)
-        )
+  internalReapplyValidatedTx = defaultReapplyValidatedTx @"LEDGER" AllegraApplyTxError
 
 instance ApplyTick AllegraEra