+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra.hs
@@ -54,22 +54,16 @@ import Cardano.Ledger.Dijkstra.TxInfo ()
 import Cardano.Ledger.Dijkstra.TxWits ()
 import Cardano.Ledger.Dijkstra.UTxO ()
 import Cardano.Ledger.Plutus (Language (..))
-import Cardano.Ledger.Rules.ValidationMode (lblStatic)
 import Cardano.Ledger.Shelley.API (
   ApplyBlock (..),
   ApplyTick (..),
   ApplyTx (..),
-  ValidatedTx (..),
-  epochInfo,
-  ruleApplyTxValidation,
-  systemStart,
+  defaultApplyTxWithValidation,
+  defaultReapplyValidatedTx,
  )
-import Cardano.Ledger.Shelley.Rules (ledgerPpL)
-import Cardano.Ledger.State (EraUTxO (..), ScriptsProvided, UTxO, utxoG)
+import Cardano.Ledger.State (EraUTxO (..), ScriptsProvided, UTxO)
 import Cardano.Slotting.EpochInfo (EpochInfo)
 import Cardano.Slotting.Time (SystemStart)
-import Control.State.Transition.Extended (ValidationPolicy (..))
-import Data.Bifunctor (Bifunctor (first))
 import Data.Foldable (toList)
 import Data.List.NonEmpty (NonEmpty)
 import qualified Data.Map.Strict as Map
@@ -85,28 +79,9 @@ instance ApplyTx DijkstraEra where
 
   mkStAnnTx = mkDijkstraStAnnTopTx
 
-  internalApplyTxWithValidation validationPolicy globals env state tx =
-    let stAnnTx =
-          mkStAnnTx
-            (epochInfo globals)
-            (systemStart globals)
-            (env ^. ledgerPpL)
-            (state ^. utxoG)
-            tx
-     in first DijkstraApplyTxError $
-          ruleApplyTxValidation @"MEMPOOL" validationPolicy globals env state stAnnTx
+  internalApplyTxWithValidation = defaultApplyTxWithValidation @"MEMPOOL" DijkstraApplyTxError
 
-  internalReapplyValidatedTx globals env state vtx =
-    fst
-      <$> first
-        DijkstraApplyTxError
-        ( ruleApplyTxValidation @"MEMPOOL"
-            (ValidateSuchThat (notElem lblStatic))
-            globals
-            env
-            state
-            (vtStAnnTx vtx)
-        )
+  internalReapplyValidatedTx = defaultReapplyValidatedTx @"MEMPOOL" DijkstraApplyTxError
 
 instance ApplyTick DijkstraEra